new 'small' & 'accordion' attribute. Cleanup of code
