+++ web/src/stories/types.ts
@@ -1,5 +1,14 @@
 export type Link = { href: string; text: string }
 
+export type Card = {
+  title: string
+  content: string
+  actions: boolean
+  small: boolean
+  accordion: boolean
+  onClick: () => void
+}
+
 export type FieldMeta = {
   label?: string
   helperText?: string
